requirements.txt
hydra-core==1.1.1
hydra-colorlog==1.1.0
pytorch-lightning==1.5.7
pandas==1.3.5
scikit-learn==1.0.2
matplotlib==3.5.1